Split mempool config into its own module
Also:
- expand config docs
- clean up mempool imports
